Added frontend UI

+++ backend/controllers/jobController.js
@@ -1,8 +1,9 @@
 import asyncHandler from "express-async-handler";
 import Job from "../models/jobSchema.js";
+import User from "../models/userSchema.js";
 
 export const getJobs = asyncHandler(async (req, res) => {
-  const jobs = await Job.find();
+  const jobs = await Job.find({ user: req.user.id });
   res.status(200).send(jobs);
 });
 
@@ -11,7 +12,7 @@ export const setJobs = asyncHandler(async (req, res) => {
     throw new Error("data is not provided");
   }
   const text = req.body.jobText;
-  const job = await Job.create({ jobText: text });
+  const job = await Job.create({ jobText: text, user: req.user.id });
   res.status(200).send(job);
 });
 
@@ -24,6 +25,15 @@ export const updateJob = asyncHandler(async (req, res) => {
   if (!job) {
     res.status(400);
     throw new Error("Job id not found");
+  }
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  if (job.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
   const updatedJob = await Job.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
@@ -40,6 +50,15 @@ export const deleteJob = asyncHandler(async (req, res) => {
   if (!job) {
     res.status(400);
     throw new Error("Job id not found");
+  }
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  if (job.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
   await job.deleteOne();
   res.status(200).send({ message: `${req.params.id} is deleted` });
